chore: code review improvements
